Componentized Activities Lifecycle
